feat: BIrdCard 스켈레톤 UI 적용
